PROGETTO1 (DAMIANO)/GestioneAssets/src/requirements.txt
flask==2.0.2
werkzeug==2.0.2
Flask-Mail
mysql-connector-python